requirements.txt
discord.py==2.1.0
openai==0.25.0
python-decouple==3.6
